+++ CHANGES.rst
@@ -28,6 +28,7 @@ Furthermore, the shell scripts ``wine-python``, ``wine-pip`` and ``wine-pytest``
 
 * FEATURE: ``wineprefix`` becomes a configuration parameter definable by users allowing custom wine prefixes, see issue #44.
 * FEATURE: Introduced new error types specific to this package.
+* FEATURE: Timeouts for start and stop of the server component can be configured.
 * FIX: *zugbruecke* did not capture and forward data coming from Windows DLLs and binaries through ``stdout`` and ``stderr``(running with Wine) most of the time.
 * FIX: ``wine-pip`` previously would, on every launch, download ``get-pip.py`` and try to install it first before running - even if ``pip`` was already installed. ``wenv pip`` does not show this behavior anymore.
 * FIX: A proper ``TimeoutError`` is raised (instead of a ``SyntaxError``) if *zugbruecke*'s server component does not start.